# apps/backend/docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/settlement/run": {
      "post": {
        "operationId": "SettlementController_run_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunSettlementDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettlementDto"
                }
              }
            }
          },
          "404": {
            "description": "Provider not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Nothing unsettled for this provider",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Run a settlement payout for a provider, covering every commission calculation not yet settled",
        "tags": [
          "Settlement"
        ]
      }
    },
    "/api/v1/settlement": {
      "get": {
        "operationId": "SettlementController_list_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SettlementDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all settlements",
        "tags": [
          "Settlement"
        ]
      }
    },
    "/api/v1/settlement/{id}": {
      "get": {
        "operationId": "SettlementController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettlementDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one settlement",
        "tags": [
          "Settlement"
        ]
      }
    },
    "/api/v1/refund/policies": {
      "post": {
        "operationId": "RefundPolicyController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRefundPolicyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundPolicyDto"
                }
              }
            }
          },
          "404": {
            "description": "Referenced category/service/provider/town-village does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "An active policy already exists for this reason at this exact scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a refund policy for a given reason at a given scope",
        "tags": [
          "Refund"
        ]
      },
      "get": {
        "operationId": "RefundPolicyController_list_v1",
        "parameters": [
          {
            "name": "scopeType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PLATFORM",
                "CATEGORY",
                "SERVICE",
                "PROVIDER",
                "GEOGRAPHY"
              ]
            }
          },
          {
            "name": "reason",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "CUSTOMER_CANCELLED",
                "PROVIDER_CANCELLED",
                "CUSTOMER_NO_SHOW",
                "PROVIDER_NO_SHOW",
                "BOOKING_REJECTED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RefundPolicyDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List refund policies, optionally filtered by scope type and/or reason",
        "tags": [
          "Refund"
        ]
      }
    },
    "/api/v1/refund/policies/{id}": {
      "get": {
        "operationId": "RefundPolicyController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundPolicyDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one refund policy",
        "tags": [
          "Refund"
        ]
      },
      "patch": {
        "operationId": "RefundPolicyController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRefundPolicyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundPolicyDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Reactivating would collide with another active policy for the same reason at the same scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a policy's refund percentage or active state",
        "tags": [
          "Refund"
        ]
      }
    },
...
    "/api/v1/refund/{id}": {
      "get": {
        "operationId": "RefundController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one refund",
        "tags": [
          "Refund"
        ]
      }
    }
  },
  "info": {
    "title": "SevaMitra API",
    "description": "SevaMitra backend API — see docs/ARCHITECTURE.md",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorDetailDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "BAD_REQUEST"
          },
          "message": {
            "type": "string",
            "example": "Incorrect OTP"
          },
          "details": {
            "type": "object"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetailDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "SettlementDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "providerId": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "totalAmount": {
            "type": "string",
            "example": "1078.25"
          },
          "currency": {
            "type": "string",
            "example": "INR"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PAID",
              "FAILED"
            ],
            "example": "PAID"
          },
          "payoutReference": {
            "type": "object",
            "nullable": true
          },
          "failureReason": {
            "type": "object",
            "nullable": true
          },
          "paidAt": {
            "type": "object",
            "nullable": true
          },
          "commissionCalculationIds": {
            "description": "Ids of the CommissionCalculation rows this settlement paid out (empty if FAILED)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "providerId",
          "totalAmount",
          "currency",
          "status",
          "commissionCalculationIds",
          "createdAt",
          "updatedAt"
        ]
      },
      "RunSettlementDto": {
        "type": "object",
        "properties": {
          "providerId": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          }
        },
        "required": [
          "providerId"
        ]
      },
      "CreateRefundPolicyDto": {
        "type": "object",
        "properties": {
          "scopeType": {
            "type": "string",
            "enum": [
              "PLATFORM",
              "CATEGORY",
              "SERVICE",
              "PROVIDER",
              "GEOGRAPHY"
            ],
            "example": "PLATFORM"
          },
          "categoryId": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
            "description": "Required when scopeType=CATEGORY"
          },
          "serviceId": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
            "description": "Required when scopeType=SERVICE"
          },
          "providerId": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
            "description": "Required when scopeType=PROVIDER"
          },
          "townVillageId": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
            "description": "Required when scopeType=GEOGRAPHY"
          },
          "reason": {
            "type": "string",
            "enum": [
              "CUSTOMER_CANCELLED",
              "PROVIDER_CANCELLED",
              "CUSTOMER_NO_SHOW",
              "PROVIDER_NO_SHOW",
              "BOOKING_REJECTED"
            ],
            "example": "CUSTOMER_CANCELLED"
          },
          "refundPercentage": {
            "type": "number",
            "example": 50,
            "description": "0 = no refund, 100 = full refund, anything between = partial"
          }
        },
        "required": [
          "scopeType",
          "reason",
          "refundPercentage"
        ]
      },
      "RefundPolicyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "scopeType": {
            "type": "string",
            "enum": [
              "PLATFORM",
              "CATEGORY",
              "SERVICE",
              "PROVIDER",
              "GEOGRAPHY"
            ],
            "example": "PLATFORM"
          },
          "categoryId": {
            "type": "object",
            "nullable": true
          },
          "serviceId": {
            "type": "object",
            "nullable": true
          },
          "providerId": {
            "type": "object",
            "nullable": true
          },
          "townVillageId": {
            "type": "object",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "enum": [
              "CUSTOMER_CANCELLED",
              "PROVIDER_CANCELLED",
              "CUSTOMER_NO_SHOW",
              "PROVIDER_NO_SHOW",
              "BOOKING_REJECTED"
            ],
            "example": "CUSTOMER_CANCELLED"
          },
          "refundPercentage": {
            "type": "string",
            "example": "50.00"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "scopeType",
          "reason",
          "refundPercentage",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateRefundPolicyDto": {
        "type": "object",
        "properties": {
          "refundPercentage": {
            "type": "number",
            "example": 75
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "RefundDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "bookingId": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "reason": {
            "type": "string",
            "enum": [
              "CUSTOMER_CANCELLED",
              "PROVIDER_CANCELLED",
              "CUSTOMER_NO_SHOW",
              "PROVIDER_NO_SHOW",
              "BOOKING_REJECTED"
            ],
            "example": "CUSTOMER_CANCELLED"
          },
          "appliedPolicyId": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "grossPaidAmount": {
            "type": "string",
            "example": "599.00"
          },
          "refundAmount": {
            "type": "string",
            "example": "299.50"
          },
          "currency": {
            "type": "string",
            "example": "INR"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "REFUNDED",
              "FAILED"
            ],
            "example": "REFUNDED"
          },
          "refundReference": {
            "type": "object",
            "nullable": true
          },
          "failureReason": {
            "type": "object",
            "nullable": true
          },
          "refundedAt": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "bookingId",
          "reason",
          "appliedPolicyId",
          "grossPaidAmount",
          "refundAmount",
          "currency",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}